+++ client/web/src/site-admin/SiteAdminPingsPage.tsx
@@ -131,10 +131,7 @@ export const SiteAdminPingsPage: React.FunctionComponent<Props> = props => {
                         <li>Search filters used (e.g. "type:", "repo:", "file:", "lang:", etc.)</li>
                     </ul>
                 </li>
-                <li>
-                    Aggregate daily, weekly, and monthly latencies (in ms) of code intelligence events (e.g., hover
-                    tooltips) and search queries
-                </li>
+                <li>Aggregate daily, weekly, and monthly latencies (in ms) of search queries</li>
                 <li>
                     Aggregate daily, weekly, and monthly counts of:
                     <ul>