+++ cmd/frontend/types/types.go
@@ -230,18 +230,8 @@ type CodeIntelEventCategoryStatistics struct {
 // to the updatecheck handler. This struct is marshalled and sent to
 // BigQuery, which requires the input match its schema exactly.
 type CodeIntelEventStatistics struct {
-	UsersCount     int32
-	EventsCount    *int32
-	EventLatencies *CodeIntelEventLatencies
-}
-
-// NOTE: DO NOT alter this struct without making a symmetric change
-// to the updatecheck handler. This struct is marshalled and sent to
-// BigQuery, which requires the input match its schema exactly.
-type CodeIntelEventLatencies struct {
-	P50 float64
-	P90 float64
-	P99 float64
+	UsersCount  int32
+	EventsCount *int32
 }
 
 // NOTE: DO NOT alter this struct without making a symmetric change